md_recommendation_system/recommender/requirements.txt
torch
torchvision
scikit-learn
numpy
pandas
tqdm
PyYAML
pymongo